답변 페이지 api수정
